Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pytest==7.4.3
httpx==0.25.2
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
xgboost==2.0.3
firebase-admin==6.2.0
slowapi==0.1.8
cryptography==41.0.7
